fix : move 'recoil store' to recoil folder
