+++ src/recoil/store.tsx
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const activeId = atom<any>({
+  key: 'activeId',
+  default: 1,
+});
